Work on deserialisation

+++ workcraft/plugins/serialisation/xml/IntSerialiser.java
@@ -11,6 +11,6 @@ public class IntSerialiser implements BasicXMLSerialiser{
 
 	public void serialise(Element element, Object object)
 			throws SerialisationException {
-		element.setAttribute("value", ((Integer)object).toString());
+		element.setAttribute("value", object.toString());
 	}
 }
